rename empty free to noopfree + suggestions
